+++ src/test/java/com/nextstep/lotto/result/RevenueRatioTest.java
@@ -1,4 +1,4 @@
-package com.nextstep.lotto;
+package com.nextstep.lotto.result;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -9,8 +9,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import com.nextstep.lotto.result.LottoMatch;
-import com.nextstep.lotto.result.RevenueRatio;
 import com.nextstep.lotto.store.LottoStore;
 
 public class RevenueRatioTest {
